+++ src/components/Button/Button.tsx
@@ -1,10 +1,13 @@
 import PropTypes from 'prop-types'
-import React, { ReactNode, CSSProperties } from 'react'
+import React, { ReactNode, CSSProperties, SyntheticEvent } from 'react'
 
 import { UIComponent, childrenExist, customPropTypes, IRenderResultConfig } from '../../lib'
 import buttonRules from './buttonRules'
 import buttonVariables from './buttonVariables'
+import Icon from '../Icon'
+import Text from '../Text'
 
+export type IconPosition = 'before' | 'after'
 export type ButtonType = 'primary' | 'secondary'
 
 export interface IButtonProps {
@@ -13,7 +16,11 @@ export interface IButtonProps {
   circular?: boolean
   className?: string
   content?: ReactNode
+  disabled?: boolean
   fluid?: boolean
+  icon?: boolean | string
+  iconPosition?: IconPosition
+  onClick?: (e: SyntheticEvent, props: IButtonProps) => void
   style?: CSSProperties
   type?: ButtonType
 }
@@ -45,11 +52,27 @@ class Button extends UIComponent<IButtonProps, any> {
     /** Additional classes. */
     className: PropTypes.string,
 
+    /** A button can show it is currently unable to be interacted with. */
+    disabled: PropTypes.bool,
+
     /** Shorthand for primary content. */
     content: customPropTypes.contentShorthand,
 
     /** A button can take the width of its container. */
     fluid: PropTypes.bool,
+
+    /** Button can have an icon. */
+    icon: customPropTypes.some([PropTypes.bool, PropTypes.string]),
+
+    /** An icon button can format an Icon to appear before or after the button */
+    iconPosition: PropTypes.oneOf(['before', 'after']),
+
+    /**
+     * Called after user's click.
+     * @param {SyntheticEvent} event - React's original SyntheticEvent.
+     * @param {object} data - All props.
+     */
+    onClick: PropTypes.func,
 
     /** A button can be formatted to show different levels of emphasis. */
     type: PropTypes.oneOf(['primary', 'secondary']),
@@ -61,7 +84,11 @@ class Button extends UIComponent<IButtonProps, any> {
     'circular',
     'className',
     'content',
+    'disabled',
     'fluid',
+    'icon',
+    'iconPosition',
+    'onClick',
     'type',
   ]
 
@@ -74,13 +101,54 @@ class Button extends UIComponent<IButtonProps, any> {
     classes,
     rest,
   }: IRenderResultConfig<IButtonProps>): ReactNode {
-    const { children, content } = this.props
+    const { children, content, disabled, icon, iconPosition, type } = this.props
+    const primary = type === 'primary'
+
+    const getContent = (): ReactNode => {
+      if (childrenExist(children)) {
+        return children
+      }
+
+      const iconIsAfterButton = iconPosition === 'after'
+      const renderedContent = [
+        content && <Text key="btn-content" truncated content={content} />,
+        icon &&
+          typeof icon === 'string' && (
+            <Icon
+              key="btn-icon"
+              name={icon}
+              xSpacing={!content ? 'none' : iconIsAfterButton ? 'before' : 'after'}
+              color={primary ? 'white' : 'black'}
+            />
+          ),
+      ].filter(Boolean)
+
+      return iconIsAfterButton ? renderedContent : renderedContent.reverse()
+    }
 
     return (
-      <ElementType {...rest} className={classes.root}>
-        {childrenExist(children) ? children : content}
+      <ElementType
+        className={classes.root}
+        disabled={disabled}
+        onClick={this.handleClick}
+        {...rest}
+      >
+        {getContent()}
       </ElementType>
     )
+  }
+
+  private handleClick = (e: SyntheticEvent) => {
+    const { onClick, disabled } = this.props
+
+    if (disabled) {
+      e.preventDefault()
+      return
+    }
+
+    if (onClick) {
+      onClick(e, this.props)
+    }
   }
 }