+++ test/specs/components/Header/HeaderDescription-test.ts
@@ -1,0 +1,7 @@
+import { isConformant } from 'test/specs/commonTests'
+
+import { HeaderDescription } from 'src/components/Header'
+
+describe('HeaderDescription', () => {
+  isConformant(HeaderDescription)
+})